Keybon12: update to Arduino code and README
This new code architecture should make it easier to both add more
screens if needed, but also to add more functions, including more
generic ones.
While usable, this is still a WIP regarding the architecture.

Also, this work on Keybon12 is backward compatible with the host
software developped by CoretechR.